+++ Mopy/bash/mod_files.py
@@ -348,7 +348,7 @@ class ModFile(object):
                 return indices[m_name] if obj_id >= 0x800 else 0
         def mapper(fid):
             if fid is None: return None
-            if isinstance(fid, (int, long)): return fid
+            if isinstance(fid, (int, int)): return fid
             modName, object_id = fid
             return (_master_index(modName, object_id) << 24) | object_id
         return mapper
